+++ mediation/Tapjoy/source/plugin/Assets/GoogleMobileAds/Platforms/Mediation/TapjoyClientFactory.cs
@@ -22,7 +22,7 @@ namespace GoogleMobileAds.Mediation
 {
     public class TapjoyClientFactory
     {
-        public static ITapjoyClient TapjoyInstance()
+        public static ITapjoyClient CreateTapjoyClient()
         {
             #if UNITY_EDITOR
             return new GoogleMobileAds.Common.Mediation.Tapjoy.DummyClient();