+++ mediation/Tapjoy/source/plugin/Assets/GoogleMobileAds/Platforms/iOS/Mediation/Tapjoy/Externs.cs
@@ -27,6 +27,9 @@ namespace GoogleMobileAds.iOS.Mediation.Tapjoy
 
         [DllImport("__Internal")]
         internal static extern void GADUMTapjoySubjectToGDPR(bool gdprApplicability);
+
+        [DllImport("__Internal")]
+        internal static extern void GADUMTapjoySetUSPrivacy(string privacyString);
     }
 }